move stores directory under structures directory
