+++ src/lib/handlers/command.handler.ts
@@ -8,9 +8,9 @@ import {
 import { yellowBright, greenBright, redBright } from 'colorette';
 import Handler from './handler';
 import Command from '../structures/pieces/command.piece';
-import CommandStore from '../stores/command.store';
+import CommandStore from '../structures/stores/command.store';
 import Guard, { GuardResult } from '../structures/pieces/guard.piece';
-import GuardStore from '../stores/guard.store';
+import GuardStore from '../structures/stores/guard.store';
 import Container from '../container';
 
 class CommandHandler implements Handler {
