+++ src/lib/handlers/event.handler.ts
@@ -2,7 +2,7 @@ import { Client, ClientEvents } from 'discord.js';
 import { redBright, yellowBright } from 'colorette';
 import Handler from './handler';
 import Event from '../structures/pieces/event.piece';
-import EventStore from '../stores/event.store';
+import EventStore from '../structures/stores/event.store';
 import Container from '../container';
 
 class EventHandler implements Handler {
