+++ src/lib/structures/stores/guard.store.ts
@@ -1,7 +1,7 @@
 import { Collection } from 'discord.js';
 import Store from './store';
-import Guard from '../structures/pieces/guard.piece';
-import PiecesLoader from '../utils/pieces/pieces-loader';
+import Guard from '../pieces/guard.piece';
+import PiecesLoader from '../../utils/pieces/pieces-loader';
 
 class GuardStore implements Store<Guard> {
   private guards!: Collection<string, Guard>;
